Refuse to run custom pipelines in the 'reform' namespace
The job artifacts pointed to by

https://source.mnt.re/reform/reform-debian-packages/-/jobs/artifacts/main/download?job=reprepro

will come from the latest CI build of the branch with the same top
commit as 'main'. This is a problem as that link is used by the script
updating https://mntre.com/reform-debian-repo. So even if a pipeline run
with custom variable settings is done on a branch other than 'main', its
results will be put behind above link, thus tainting the repository.

To prevent this, forbid any pipeline runs with custom variable values in
the 'reform' namespace. To do custom runs, fork reform-debian-packages
under your own username instead.