build_patched.sh: check whether patch is newer than last version in repository
Previously, packages would not get rebuilt if the cached version in the
local repo was greater than the version in Debian. Since we make the
version in our repo greater than the version in Debian by adding a date
suffix, this results in packages not getting rebuilt even if their
patches change. Thus, compare the timestamp of the patch with the
timestamp encoded in the package version in the repository.

Thanks: Æsthr